Make shared code separate
The work is partial, and does not compile.
